+++ src/Components/Homepage/Homepage.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import './Homepage.css';
-import SimpleMap from './map';
 
 export default class Homepage extends React.Component {  
     constructor(props) {
@@ -22,10 +21,6 @@ export default class Homepage extends React.Component {
         event.preventDefault()
         var thing = this.state.thing;
         alert(thing)
-        //need to work on this
-        //findthatshop(thing, (response) => {
-        //    self.setState({shops: response.data.shop.shops})
-        //})
     };
 
     render() {
@@ -35,7 +30,7 @@ export default class Homepage extends React.Component {
                     <h1 className="font-size-3-5 font-color-white padding-top-175px">Find That Shop</h1>
                     <form onSubmit={this.onSubmit}>
                         <div className="input-group width-65 ">
-                            <input type="text" name="thing" className="form-control" placeholder="where I can find a..." onChange={this.onChange}/>
+                            <input type="text" name="thing" className="form-control" placeholder="where I can find a..." onChange={this.onChange} />
                             <span className="input-group-btn">
                                 <button type="submit" className="btn btn-primary">Go</button>
                             </span>
@@ -43,5 +38,6 @@ export default class Homepage extends React.Component {
                     </form>
                 </div>
             </div>
-        )};
+        )
+    };
 }
