+++ src/Components/Logout/Logout.js
@@ -1,6 +1,6 @@
-import React , {Component} from 'react';
+import React from 'react';
 import { connect } from 'react-redux';
-import { logoutaction } from './action/log'
+import { logoutaction } from '../../utils/action/log'
 
 class Logout extends React.Component {
