--- src/shop.css
@@ -1,15 +1,0 @@
-.align-webkit-center {
-    text-align: webkit-center;
-}
-
-.jumbotron {
-    padding: 2rem 1rem;
-}
-
-.font-size-3-5rem {
-    font-size: 3.5rem;
-}
-
-.margin-30px {
-    margin : 30px;
-}
