+++ src/utils/store/store.js
@@ -1,6 +1,6 @@
 import { createStore, applyMiddleware } from 'redux';
 import logger from 'redux-logger'
-import IsAuth from './reducer/IsAuth';
+import IsAuth from '../reducer/IsAuth';
 
 export default createStore(
     IsAuth,
